app with endpoints

+++ src/products/products.controller.ts
@@ -1,6 +1,8 @@
-import { Controller, Get, Delete, Post, Param, Body } from '@nestjs/common';
+import { Controller, Get, Delete, Post, Put, Param, Body, NotFoundException } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { UpdateProductDTO } from './dtos/update-product.dto';
+import { ParseUUIDPipe } from '@nestjs/common';
 
 
 @Controller('products')
@@ -11,24 +13,40 @@ export class ProductsController {
 
     @Get('/')
     getAll(): any {
-      return this.productsService.getAll();
+        return this.productsService.getAll();
     }
 
     @Get('/:id')
-    getById(@Param('id') id: string) {
-      return this.productsService.getById(id);
-    } 
+    getById(@Param('id', new ParseUUIDPipe()) id: string) {
+        const product = this.productsService.getById(id);
+        if (!product) throw new NotFoundException('Product not found');
+        return product;
+    }
 
     @Post('/')
     create(@Body() productData: CreateProductDTO) {
-      return this.productsService.create(productData);
-  }   
-    
+        return this.productsService.create(productData);
+    }
+
+    @Put('/:id')
+    update(
+        @Param('id', new ParseUUIDPipe()) id: string,
+        @Body() productData: UpdateProductDTO,
+    ) {
+        if (!this.productsService.getById(id))
+            throw new NotFoundException('Product not found');
+
+        this.productsService.updateById(id, productData);
+        return { success: true };
+    }
+
     @Delete('/:id')
-    deleteById(@Param('id') id: string) {
+    deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+      if (!this.productsService.getById(id))
+        throw new NotFoundException('Product not found');
       this.productsService.deleteById(id);
       return { success: true };
-    }   
-    
+    }    
+
 
 }
